Update SiloName and pod identification in Kubernetes hosting
This commit includes several changes to improve the identification of pods and silos in Kubernetes hosting. The `SiloName` in `Configure(SiloOptions options)` method and the `clusterPods` HashSet in `OnStart(CancellationToken cancellation)` method are now set as a combination of `Namespace` and `PodName` instead of just `PodName`. A new list `managedPods` is created to filter the list of known silo pods to only the namespace of the current silo. The `podName` is now extracted from `change.Name` by splitting it on `_` and used in log information and the `DeleteNamespacedPodAsync` method. Lastly, the `TryMatchSilo(V1Pod pod, out ClusterMember server)` method now compares `member.Value.Name` with a combination of `pod.Metadata.Namespace` and `pod.Metadata.Name`.